Split tests and add parallelization
